Make code-block margins same as print blocks
The current theme adds a 24px margin to the bottom of all styles except for code-highlight blocks. This makes for an awkward, inconsistent lack-of-space below pre-formatted text.

This has the codehilite class the same as the rest of the documents.

This is the docs repo equivalent to https://github.com/opensciencegrid/technology/pull/339.