Overhaul site planning document.
Focus currently is deleting absolutely irrelevant / incorrect content
and including as many tidbits about OSG as possible.

The result is not pretty - it looks a bit like a wall of words.
Diagrams need to be added back, as well as some tables to help illustrate
the decision tree.

Have to start somewhere!