Remove problematic filenames for Mac OS X filesystems.
Have pity on an old guy: Mac OS X filesystems have historically been
case insensitive, which causes all sort of havoc with git when trying
to manage the docs repo on a Mac laptop.

This cleans up the problem by simply renaming one of the two files that
git considers identical; it's in the archive, so I don't think there's much
risk in this.