Scrub extra references to bestman2.
This commit removes a few unnecessary references to bestman2; I've
left all the documentation itself, but just made the HDFS and Xrootd
docs a bit less bestman2-centric.

Additionally, this includes a mild cleanup of the Hadoop docs.  Honestly,
they are still fairly low-quality - but can be further improved over
a series of updates.