reword "Applicable versions" section (SOFTWARE-3075)
"OSG Version" presumably refers to the nebulous version of the
not-especially-useful osg-version package.  What is really important
is the release series that the other osg packages was installed from,
if i understand correctly.

Also, the sentence about upgrading to a 3.5 version does not state
explicitly that it is referring to frontier-squid (as opposed to the
"OSG Version" or something); and, while that can be inferred by a
careful reading of the rest of the paragraph, it would make it more
clear / obvious to mention explicitly in this sentence that we are
talking about frontier-squid.