Remove LBL links for bestman2 docs.
The LBL links for bestman2 are periodically broken and increasingly
less relevant, especially as we near the end of the support.

I don't see much value in including these as a reference (after all,
we don't expect to have any new bestman2 services in the next 2.5 months),
they can always be found via Google, and they have a noticable cost
because of how often they break the CI tests.