Begin overhaul of the cache install.
This addresses major readability and content issues with the cache
installation document.

It's not obvious one could install either authenticated or public
caches using this documentation.  However, it's a start.

I left several TODO items in the document where it instructed the
human to perform steps that obviously are part of missing packaging.
The StashCache packaging should be fixed and these TODOs eliminated
completely.