Add most of the information from osgvo-docker-pilot README
It's been reformatted to the OSG docs style, cleaned up and reorganized
a bit been cleaned up, and I put the cvmfsexec example at the top
instead of the bind-mount example.

If this looks good, I will replace most of osgvo-docker-pilot README
with a link to this page, except for the [Singularity/BYOR section]
(https://github.com/opensciencegrid/osgvo-docker-pilot#singularity--bring-your-own-resources).